Adding tests to demonstrate faulty eventReceive callback instead of eventDrop during rerender while dragging. Refs fullcalendar/fullcalendar#6747
